+++ NotationFileManager.h
@@ -38,7 +38,7 @@ OSStatus CreateTemporaryFile(FSRef *parentRef, FSRef *childTempRef);
 OSStatus CreateDirectoryIfNotPresent(FSRef *parentRef, CFStringRef subDirectoryName, FSRef *childRef);
 CFUUIDRef CopyHFSVolumeUUIDForMount(const char *mntonname);
 long BlockSizeForNotation(NotationController *controller);
-UInt32 diskUUIDIndexForNotation(NotationController *controller);
+NSUInteger diskUUIDIndexForNotation(NotationController *controller);
 
 - (void)purgeOldPerDiskInfoFromNotes;
 - (void)initializeDiskUUIDIfNecessary;
