+++ PassphraseRetriever.m
@@ -39,7 +39,7 @@
 }
 
 //1 for OK, 0 for cancelled, some other number for something else
-- (int)loadedUserPassphraseData {
+- (NSInteger)loadedUserPassphraseData {
 	
 	if (!window) {
 		if (![NSBundle loadNibNamed:@"PassphraseRetriever" owner:self])  {
@@ -71,7 +71,7 @@
 
 	[rememberKeychainButton setState:[notationPrefs storesPasswordInKeychain]];
 	
-	int result = [NSApp runModalForWindow:window];
+	NSInteger result = [NSApp runModalForWindow:window];
 	
 	[passphraseField setStringValue:@""];
 	[self textDidChange:nil];
