+++ PreviewController.m
@@ -549,7 +549,7 @@
 
 -(IBAction)switchTabs:(id)sender
 {
-	int tabSelection = [tabView indexOfTabViewItem:[tabView selectedTabViewItem]];
+	NSUInteger tabSelection = [tabView indexOfTabViewItem:[tabView selectedTabViewItem]];
 
 	if (tabSelection == 0) {
 		[tabSwitcher setTitle:@"View Preview"];
